requirements.txt
django==3.0.6
django-tinymce==3.3.0
